+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/generation/variables/ArrayStore.java
@@ -41,5 +41,6 @@ public interface ArrayStore extends ValueStore {
         INTERNAL,
         EXTERNAL,
         REMOTE,
+        REMOTE_SHARED,
     }
 }