+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/generation/variables/InternalArray.java
@@ -15,6 +15,7 @@ import org.jspecify.annotations.NullMarked;
 import org.jspecify.annotations.Nullable;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.function.Consumer;
 import java.util.stream.IntStream;
 
@@ -24,52 +25,63 @@ import static info.teksol.mc.mindcode.logic.instructions.ArrayOrganization.INTER
 @NullMarked
 public class InternalArray extends AbstractArrayStore {
     private final boolean remote;
+    private final boolean shared;
     private final int startOffset;
+    private final @Nullable LogicVariable processor;      // Actual processor in case of shared arrays
     private final LogicArray logicArray;
 
-    private InternalArray(SourcePosition sourcePosition, String name, List<ValueStore> elements, boolean remote) {
+    private InternalArray(SourcePosition sourcePosition, String name, @Nullable LogicVariable processor,
+            List<ValueStore> elements, boolean remote, boolean shared) {
         super(sourcePosition, name, elements);
+        this.processor = processor;
         this.remote = remote;
+        this.shared = shared;
         startOffset = 0;
         logicArray = LogicArray.create(this);
     }
 
-    private InternalArray(SourcePosition sourcePosition, String name, int offset, List<ValueStore> elements, boolean remote) {
+    private InternalArray(SourcePosition sourcePosition, String name, @Nullable LogicVariable processor, int offset,
+            List<ValueStore> elements, boolean remote, boolean shared) {
         super(sourcePosition, name, elements);
+        this.processor = processor;
         this.remote = remote;
+        this.shared = shared;
         startOffset = offset;
         logicArray = LogicArray.create(this, offset, offset + elements.size());
     }
 
-    public static InternalArray create(NameCreator nameCreator, AstIdentifier identifier, int size, boolean isVolatile, @Nullable LogicVariable processor) {
+    public static InternalArray create(NameCreator nameCreator, AstIdentifier identifier, int size, boolean isVolatile,
+            @Nullable LogicVariable processor, boolean shared) {
         if (processor != null) {
             CodeAssembler assembler = MindcodeCompiler.getContext().assembler();
             return new InternalArray(identifier.sourcePosition(),
-                    nameCreator.arrayBase(processor.getName(), identifier.getName()),
+                    nameCreator.arrayBase(shared ? "" : processor.getName(), identifier.getName()),
+                    shared ? processor : null,
                     IntStream.range(0, size)
                             .mapToObj(index -> (ValueStore) new RemoteVariable(identifier.sourcePosition(), processor,
                                     processor.getName() + "." + identifier.getName() + "[" + index + "]",
                                     LogicString.create(nameCreator.remoteArrayElement(identifier.getName(), index)),
-                                    assembler.nextTemp(), false, false)).toList(), true);
+                                    assembler.nextTemp(), false, false)).toList(), true, shared);
         } else {
             return new InternalArray(identifier.sourcePosition(),
-                    nameCreator.arrayBase("", identifier.getName()),
+                    nameCreator.arrayBase("", identifier.getName()), null,
                     IntStream.range(0, size)
                             .mapToObj(index -> (ValueStore) LogicVariable.arrayElement(identifier, index,
-                                    nameCreator.arrayElement(identifier.getName(), index), isVolatile)).toList(), false);
+                                    nameCreator.arrayElement(identifier.getName(), index), isVolatile)).toList(), false, false);
         }
     }
 
     public static InternalArray createConst(NameCreator nameCreator, AstIdentifier identifier, int size, List<ValueStore> elements) {
         List<ValueStore> wrappedElements = elements.stream().map(InternalArray::constantWrap).toList();
         return new InternalArray(identifier.sourcePosition(),
-                nameCreator.arrayBase("", identifier.getName()), wrappedElements, false);
+                nameCreator.arrayBase("", identifier.getName()),
+                null, wrappedElements, false, false);
     }
 
     public static InternalArray createInvalid(NameCreator nameCreator, AstIdentifier identifier, int size) {
         return new InternalArray(identifier.sourcePosition(),
-                nameCreator.arrayBase("", identifier.getName()),
-                IntStream.of(size).mapToObj(index -> (ValueStore) LogicVariable.INVALID).toList(), false);
+                nameCreator.arrayBase("", identifier.getName()), null,
+                IntStream.of(size).mapToObj(index -> (ValueStore) LogicVariable.INVALID).toList(), false, false);
     }
 
     public LogicArray getLogicArray() {
@@ -78,7 +90,7 @@ public class InternalArray extends AbstractArrayStore {
 
     @Override
     public ArrayType getArrayType() {
-        return remote ? ArrayType.REMOTE : ArrayType.INTERNAL;
+        return remote ? shared ? ArrayType.REMOTE_SHARED : ArrayType.REMOTE : ArrayType.INTERNAL;
     }
 
     @Override
@@ -88,7 +100,11 @@ public class InternalArray extends AbstractArrayStore {
 
     @Override
     public ArrayStore subarray(SourcePosition sourcePosition, int start, int end) {
-        return new InternalArray(sourcePosition, name, startOffset + start, elements.subList(start, end), remote);
+        return new InternalArray(sourcePosition, name, processor, startOffset + start, elements.subList(start, end), remote, shared);
+    }
+
+    public LogicVariable getProcessor() {
+        return Objects.requireNonNull(processor);
     }
 
     @Override
@@ -99,7 +115,7 @@ public class InternalArray extends AbstractArrayStore {
 
     @Override
     public InternalArray withSourcePosition(SourcePosition sourcePosition) {
-        return new InternalArray(sourcePosition, name, elements, remote);
+        return new InternalArray(sourcePosition, name, processor, elements, remote, shared);
     }
 
     public class InternalArrayElement implements ValueStore {